Refactor plugins directory to allow for README.md for each
