+++ examples/run_discordbot_hello_world_and_weather.js
@@ -1,10 +1,10 @@
-const HelloWorldPlugin = require('../src/plugins/hello_world.js');
-const WeatherPlugin = require('../src/plugins/weather.js');
+const HelloWorldPlugin = require('../src/plugins/hello_world/plugin.js');
+const WeatherPlugin = require('../src/plugins/weather/plugin.js');
 const DiscordBot = require('../src/discordbot.js');
 
 let plugins = [
   new HelloWorldPlugin,
-  // User can now give Weather an api 
+  // User can now give Weather an api
   new WeatherPlugin({
     openweather_api_key: process.env.OPENWEATHER_API_KEY
   })
