+++ src/plugins/hello_world/README.md
@@ -1,0 +1,2 @@
+# discordbot-hello-world
+TODO: Add information about setting up the plugin
