+++ src/plugins/poll/README.md
@@ -1,0 +1,6 @@
+# discord-poll
+
+## Todo
+* Finish implementing all methods
+* Refactor to work with plugin architecture
+* Write a test suite
