+++ src/plugins/poll/plugin.js
