+++ src/plugins/todo/README.md
@@ -1,0 +1,8 @@
+# discordbot-todo
+
+## Todo
+* Fix the plugin so that it actually works
+* Create test suite
+* Allow user to set dynamodb host so a local dynamodb client may be used
+* Allow user to set the AWS region
+* Allow the user to pass in AWS credentials, including a session token (AWS STS)
