+++ src/plugins/weather/README.md
@@ -1,0 +1,2 @@
+# discordbot-weather
+TODO: Add information about setting up and configuring this plugin
